refactor(core): extract shared image downscale + sharp loader
T1.2: prepareImageAttachment/cleanupPreparedAttachment move from the assets
enrichment engine to @bakin/core/media/downscale so chat attachments (T6.2)
and enrichment share ONE implementation of the 2 MB inline-cap downscale.

The lazy sharp loader also moves to core (@bakin/core/media/sharp-loader),
killing the documented deliberate duplicate in the images plugin — plugin
boundaries forced the copy; core doesn't. asset-media re-exports it for its
plugin-local consumers. Behavior unchanged (pinned by the moved tests).
